Update patch set 5
Patch Set 5:

I will not score this-- in my (not so humble, I guess) opinion, assert is not a solution for function argument validation in a library that produces binaries built with NDEBUG defined. 

My suggested fix is a negation of each assertion condition within a conditional immediately after the assertion. That along with a known status code for "invalid argument to API method" would be good enough, I think. 

The muxer code does the above.

Patch-set: 5