Update patch set 2
Patch Set 2:

It would probably be more efficient to replace Makefile and the MSVC stuff with one .gyp file...

Not that I want to do that right this minute, it's just that make and I don't get along. ;)

Anyway, PTAL/comments addressed excepting that the GNU dependency generation example doesn't seem to work completely.

Patch-set: 2