muxer: Do not write TrackType unless specified
Adding a generic track using Segment::AddTrack() and not specifying the track's
type results in writing of TrackType element with value 0. It is an invalid
file as per Matroska Specification [1]. Since it's an optional element, only
write if it is explicitly specified.

[1] http://www.matroska.org/technical/specs/index.html#TrackType

Fixes Issue #769: https://code.google.com/p/webm/issues/detail?id=769

Change-Id: I4d75d169fd96c7e1fad278561f0e7a3e1135989c