Update patch set 1
Patch Set 1: Do not submit

I misread the spec. Turns out this is not the correct fix. TrackType element is mandatory and not optional. Nevertheless, the bug is still there because we are defaulting to 0 which results in an invalid matroska file. According to the specification, there is no default value for the TrackType element. 

So ideally, Segment::AddTrack() should take track_type as one of its mandatory arguments. But since we cannot change that (it's a public API), we need to outright reject muxing files if Segment::AddTrack() is called but Track::set_type() is never called.

Patch-set: 1
Label: Code-Review=-2
Label: Commit-Queue=0
Label: Verified=0