Update patch set 2
Patch Set 2:

(9 comments)

Will you be using the tagging support in some project? I don't want to add this to libwebm just for the sake of having it, but if it's going to be used I have no strenuous objections.

First review pass. Style stuff, mainly. Please note the general comment in the parsing code.

Patch-set: 2
Label: Code-Review=0