Add mechanism to end Clusters with BlockDuration for all tracks.
This helps browsers implementing MediaSourceExtensions (MSE) to know
the duration of a Cluster. MSE allows users to append clusters one
at a time, and in any order. It may often arise that MSE cannot
derive duration for the final block in a cluster when later clusters
are not yet (if ever) appended.

Change-Id: I75dcb7d916296a0f3110d93209ea1bdaeab529d4