Update patch set 1
Patch Set 1:

I have tried not to change the public API as much as i can.

The only public behavior that has been changed is that Cluster::Finalize() will now always be called irrespective of kLive or kFile. Instead, we look at the IMkvWriter::Seekable() to determine whether or not to update the Cluster size.

I've tested it extensively with different kinds of files. mkvinfo on the test files will also tell you that this is WAI. Please let me know if i'm breaking some aspect of the library.

Patch-set: 1