Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > > > Perhaps another pass to default our DocTypeReadVersion to 4 as
>  > > > well. Since we don't have any reliable method of avoiding
>  > > > misbehavior it's probably better to err on the side of caution
>  > > and
>  > > > mark all mkvmuxer output as v4 for both version fields.
>  > >
>  > > Like i said before, writing v4 for DocTypeReadVersion is NOT a
>  > good
>  > > idea. Both FFmpeg and ExoPlayer outright reject such files (which
>  > > means we are breaking YT). So let's leave that at 2 for now.
>  > 
>  > Sorry, I missed that comment and was just going by the description
>  > given offline
> 
> This change is fine as it only updates DocTypeVersion.
> 
> My comment was a reply to Tom's comment where he said he was going to do another pass that would change DocTypeReadVersion.

Hooray for writing broken files because the projects that read them are broken! \o/

OK, that's a little pedantic. What we should probably do is submit patches to exoplayer and ffmpeg since it appears we're all wrong (at present, based on current spec wording). Maybe the wording changed recently. Anyway, a value of 4 for both sounds like it is strictly correct based on the current text. (especially if we're muxing opus and/or colour)

Huge meh, anyway. I'm satisfied with changing only DocTypeVersion. We have tools of our own that will get mad if DocTypeReadVersion changes (mediafoundation and dshow, probably some chrome things in addition to ffmpeg; mse for ex)

Patch-set: 6
Label: Code-Review=0