webm_parser: add OnElementEnd callback
Appending a new EBML header and Segment while parsing an existing
Segment will cause libwebm to treat the incoming EBML and Segment as
"child" of the existing Segment and bail when the new Segment's size
overflows the previous segment. This callback allows the case to be
detected and the parser reinitialized.

See the related bug:
https://bugs.webkit.org/show_bug.cgi?id=218149

Change-Id: Idc60feedf25ecddf7a802dcd0239f40757e30a6a