added SetPosition() function to Reader
The example implementations of Reader all track the stream position
internally, but they only ever increase it. When seeking it is helpful
to be able to update the internal position tracking to the new
location. For example if you were seeking backwards with the FileReader
you would want to set the private position_ variable to the new file
position because without this function it would just keep increasing
from where it was before you seeked. An alternate approach for a
similar result would be to change the position_ and pos_ private
variables to be protected instead so that the different Reader classes
could be subclassed by users to add this same functionality without
reimplementing the whole reader.

Change-Id: Iae299a3b36a8440da941b182c8865429ee3515d4